+++ core/src/main/scala/dev/profunktor/tracer/auth/AuthTracerDsl.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer
+package dev.profunktor.tracer.auth
 
-import cats.data.Kleisli
-import Tracer.TraceId
+import dev.profunktor.tracer.Tracer.TraceId
+import dev.profunktor.tracer.auth.AuthTracedHttpRoute.AuthTracedRequest
+import org.http4s.AuthedRequest
 
-object Trace {
-  type Trace[F[_], A] = Kleisli[F, TraceId, A]
-
-  def apply[F[_], A](run: TraceId => F[A]): Trace[F, A] = Kleisli[F, TraceId, A](run)
+trait AuthTracerDsl {
+  object using {
+    def unapply[T, F[_]](tr: AuthTracedRequest[F, T]): Option[(AuthedRequest[F, T], TraceId)] =
+      Some(tr.request -> tr.traceId)
+  }
 }
