+++ core/src/main/scala/dev/profunktor/tracer/auth/Http4sAuthTracerDsl.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer
+package dev.profunktor.tracer.auth
 
 import org.http4s.dsl.Http4sDsl
 
-trait Http4sTracerDsl[F[_]] extends Http4sDsl[F] with TracerDsl
+trait Http4sAuthTracerDsl[F[_]] extends Http4sDsl[F] with AuthTracerDsl
 
-object Http4sTracerDsl {
-  def apply[F[_]]: Http4sTracerDsl[F] = new Http4sTracerDsl[F] {}
+object Http4sAuthTracerDsl {
+  def apply[F[_]]: Http4sAuthTracerDsl[F] = new Http4sAuthTracerDsl[F] {}
 }
