+++ core/src/main/scala/dev/profunktor/tracer/instances/tracer.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer
+package dev.profunktor.tracer
+package instances
 
-import cats.effect.IO
-import cats.syntax.functor._
-
-object IOAssertion {
-  def apply[A](ioa: IO[A]): Unit = ioa.void.unsafeRunSync
+object tracer {
+  implicit def defaultTracer[F[_]]: Tracer[F] = Tracer.create[F]()
 }
