+++ core/src/test/scala/dev/profunktor/tracer/IOAssertion.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer
+package dev.profunktor.tracer
 
-import scala.reflect.ClassTag
+import cats.effect.IO
+import cats.syntax.functor._
 
-trait TracerLog[F[_]] {
-  def info[A: ClassTag](value: => String): F[Unit]
-  def error[A: ClassTag](value: => String): F[Unit]
-  def warn[A: ClassTag](value: => String): F[Unit]
+object IOAssertion {
+  def apply[A](ioa: IO[A]): Unit = ioa.void.unsafeRunSync
 }
