+++ examples/src/main/scala/com/github/gvolpe/tracer/Main.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer
+package dev.profunktor.tracer
 
 import cats.effect._
 import cats.syntax.all._
 import cats.temp.par._
-import com.github.gvolpe.tracer.Trace.Trace
-import com.github.gvolpe.tracer.module._
-import com.github.gvolpe.tracer.module.tracer._
+import dev.profunktor.tracer.Trace.Trace
+import dev.profunktor.tracer.module._
+import dev.profunktor.tracer.module.tracer._
 import org.http4s.client.blaze.BlazeClientBuilder
 import org.http4s.server.blaze.BlazeServerBuilder
