+++ examples/src/main/scala/com/github/gvolpe/tracer/Server.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer
+package dev.profunktor.tracer
 
 import cats.effect._
 import cats.syntax.functor._
-import com.github.gvolpe.tracer.instances.tracerlog._
+import dev.profunktor.tracer.instances.tracerlog._
 
 object Server extends IOApp {
 
-  // For a default instance with header name "Trace-Id" just use `import com.github.gvolpe.tracer.instances.tracer._`
+  // For a default instance with header name "Trace-Id" just use `import dev.profunktor.tracer.instances.tracer._`
   implicit val tracer = Tracer.create[IO]("Flow-Id")
 
   override def run(args: List[String]): IO[ExitCode] =
