+++ examples/src/main/scala/com/github/gvolpe/tracer/algebra/UserAlgebra.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer.algebra
+package dev.profunktor.tracer.algebra
 
-import com.github.gvolpe.tracer.model.user.{User, Username}
+import dev.profunktor.tracer.model.user.{User, Username}
 
 trait UserAlgebra[F[_]] {
   def find(username: Username): F[User]
