+++ examples/src/main/scala/com/github/gvolpe/tracer/model/user.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer.model
+package dev.profunktor.tracer.model
 
 object user {
   final case class Username(value: String) extends AnyVal
