+++ examples/src/main/scala/com/github/gvolpe/tracer/module/HttpClients.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer.module
+package dev.profunktor.tracer.module
 
 import cats.effect.Sync
-import com.github.gvolpe.tracer.http.client.{LiveUserRegistry, UserRegistry}
+import dev.profunktor.tracer.http.client.{LiveUserRegistry, UserRegistry}
 import org.http4s.client.Client
 
 private[module] trait HttpClients[F[_]] {
