+++ examples/src/main/scala/com/github/gvolpe/tracer/module/Programs.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer.module
+package dev.profunktor.tracer.module
 
 import cats.effect.Sync
 import cats.temp.par._
-import com.github.gvolpe.tracer.algebra.UserAlgebra
-import com.github.gvolpe.tracer.program.UserProgram
+import dev.profunktor.tracer.algebra.UserAlgebra
+import dev.profunktor.tracer.program.UserProgram
 
 private[module] trait Programs[F[_]] {
   def users: UserAlgebra[F]
