+++ examples/src/test/scala/com/github/gvolpe/tracer/http/UserRoutesSpec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer.http
+package dev.profunktor.tracer.http
 
 import cats.effect.{ContextShift, IO}
-import com.github.gvolpe.tracer.Trace._
-import com.github.gvolpe.tracer.http.client.TestUserRegistry
-import com.github.gvolpe.tracer.instances.tracer._
-import com.github.gvolpe.tracer.model.user.{User, Username}
-import com.github.gvolpe.tracer.program.UserProgram
-import com.github.gvolpe.tracer.repository.TestUserRepository
+import dev.profunktor.tracer.Trace._
+import dev.profunktor.tracer.http.client.TestUserRegistry
+import dev.profunktor.tracer.instances.tracer._
+import dev.profunktor.tracer.model.user.{User, Username}
+import dev.profunktor.tracer.program.UserProgram
+import dev.profunktor.tracer.repository.TestUserRepository
 import org.http4s.Method._
 import org.http4s.{Request, Status, Uri}
 import org.scalatest.prop.TableFor3
