+++ examples/src/test/scala/com/github/gvolpe/tracer/program/UserProgramSpec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer.program
+package dev.profunktor.tracer.program
 
 import cats.effect.{ContextShift, IO}
-import com.github.gvolpe.tracer.IOAssertion
-import com.github.gvolpe.tracer.http.client.TestUserRegistry
-import com.github.gvolpe.tracer.model.user.{User, Username}
-import com.github.gvolpe.tracer.repository.TestUserRepository
+import dev.profunktor.tracer.IOAssertion
+import dev.profunktor.tracer.http.client.TestUserRegistry
+import dev.profunktor.tracer.model.user.{User, Username}
+import dev.profunktor.tracer.repository.TestUserRepository
 import org.scalatest.FunSuite
 
 import scala.concurrent.ExecutionContext
