+++ log4cats/src/main/scala/com/github/gvolpe/tracer/log4cats.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2019 Gabriel Volpe
+ * Copyright 2018-2019 ProfunKtor
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.github.gvolpe.tracer
+package dev.profunktor.tracer
 
-import com.github.gvolpe.tracer.Trace._
+import dev.profunktor.tracer.Trace._
 import io.chrisdavenport.log4cats.Logger
 
 import scala.reflect.ClassTag
